+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 1.5 
 - Reinvest benefit parameter #27
 - Don't estimate buy, buy at lowest bid on the order book #31
+- - New algos 8,9 and algo migration for 4->7 to 8 and 9.
 - Buy order auto cancellation #14
-- minimum_profit=1.002
+- minimum_profit=1.005
 
 # 1.4
 - Only create orders above or upper the current price #17